move <Position|Selector> → sys.ui.common
