+++ code/system.ui/react.common/src/test.ui/entry.Specs.mts
@@ -16,6 +16,8 @@ export const ModuleSpecs = {
   'sys.ui.common.LoadPanel': () => import('../ui/LoadPanel/-SPEC'),
   'sys.ui.common.Measure': () => import('../ui.tools/Measure/-SPEC'),
   'sys.ui.common.RenderCount': () => import('../ui/RenderCount/-SPEC'),
+  'sys.ui.common.Position': () => import('../ui/Position/-SPEC'),
+  'sys.ui.common.Position.Selector': () => import('../ui/Position.Selector/-SPEC'),
   'sys.ui.common.ProgressBar': () => import('../ui/ProgressBar/-SPEC'),
   'sys.ui.common.PropList': () => import('../ui/PropList/-dev/-SPEC'),
   'sys.ui.common.PropList.FieldSelector': () => import('../ui/PropList.FieldSelector/-dev/-SPEC'),
