+++ code/system.ui/react.common/src/ui/Position.Selector/common.ts
@@ -1,15 +1,15 @@
 import { type t } from './common';
+
 export * from '../common';
+export { Grid } from '../Grid';
 
 const selected: t.Pos = ['left', 'top'];
-const video: t.ConceptSlugVideo = {};
 
 /**
  * Constants
  */
 export const DEFAULTS = {
   enabled: true,
-  video,
   size: 150,
   selected,
 } as const;
