+++ code/system.ui/react.common/src/ui/Position/Wrangle.mts
@@ -1,5 +1,5 @@
 import { DEFAULTS, type t } from './common';
-import { Wrangle as WrangleSelector } from '../ui.PositionSelector/Wrangle.mjs';
+import { Wrangle as WrangleSelector } from '../Position.Selector/Wrangle.mjs';
 
 const { pos } = WrangleSelector;
